+++ lisa-topology/src/main/scala/lisa/Main.scala
@@ -1,12 +1,3 @@
 package lisa
 
-import lisa.TopologyLibrary
-import lisa.Topology
-import lisa.prooflib.BasicMain
-
-trait Main extends BasicMain {
-
-  export TopologyLibrary.{given, _}
-  export Topology.{given, _}
-
-}
+import lisa.maths.topology
